refactor: Move styles builder to new files
